export post, update, delete functionality in services
