+++ src/pages/admin/Zones.tsx
@@ -6,15 +6,13 @@ import { SubmitHandler } from 'react-hook-form'
 import { NewZoneFormData, newZoneSchema } from '@/schemas/newZoneSchema'
 import ZonesTable from '@/components/features/admin/ZonesTable'
 import NewZoneForm from '@/components/features/forms/NewZoneForm'
-import { useSnackbar } from '@/hooks/useSnackbar'
-import { useMutation, useQueryClient } from '@tanstack/react-query'
-import { getWarehouseManagementApi } from '@/services/generated-api'
+import usePostZone from '@/hooks/services/zones/usePostZone'
 
 export default function Zones() {
   const { t: translate } = useTranslation()
   const [openDialog, setOpenDialog] = useState(false)
   const [zoneName, setZoneName] = useState('')
-  const { showSnackbar } = useSnackbar()
+  const mutationPost = usePostZone(zoneName)
 
   const handleClickOpen = () => {
     setOpenDialog(true)
@@ -23,25 +21,6 @@ export default function Zones() {
   const onCloseDialog = () => {
     setOpenDialog(false)
   }
-
-  const queryClient = useQueryClient()
-
-  const mutationPost = useMutation({
-    mutationFn: getWarehouseManagementApi().postApiZoneAdd,
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: ['zones'] })
-      showSnackbar({
-        message: translate('newZone.snackBar.messages.createZone.success', { name: zoneName }),
-        type: 'success'
-      })
-    },
-    onError: () => {
-      showSnackbar({
-        message: translate('newZone.snackBar.messages.createZone.error'),
-        type: 'error'
-      })
-    }
-  })
 
   const handleSubmit: SubmitHandler<NewZoneFormData> = (data) => {
     setZoneName(data.zoneName)
